gianfar: Add debugfs packet forwarding loopback
Add the GFAR_DBG_LOOP config option and debugfs support
to be able to configure packet forwarding loops in the
Gianfar driver.  This config option is disabled by default
and should be used only for debugging/ benchmarking
purposes.
For details on how to configure a packet forwarding debug
loop refer to the help section of CONFIG_GFAR_DBG_LOOP.

Change-Id: I24d440edc4c7d63a055205e022b743e30f58409b
Signed-off-by: Claudiu Manoil <claudiu.manoil@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22861
Reviewed-by: Richard Schmitt <richard.schmitt@freescale.com>
Tested-by: Richard Schmitt <richard.schmitt@freescale.com>