dpaa_eth: merge recycle and confirmation queues
In FMan v3, special transmission queues, called recycle queues, were created
because it was believed that the FCO flag from the FD is not enough to
differentiate between frames that are recyclable and those who are not. The
frame queue flags were used to signal the BMI to release the buffers back to
the default buffer pool.

The last findings and FMan improvements permit recycling activation only by
using FD fields. With this patch, the recycling frame queues are not needed
anymore.

This patch also removes the code dependency on the FMan version and improves
TCP traffic in forwarding scenarios due to single transmission queues, reducing
the out-of-order segments.

Signed-off-by: Marian Rotariu <marian.rotariu@freescale.com>
Change-Id: Idd4fea4ef7883563df67cf21dea9e87435b2dee8
Reviewed-on: http://git.am.freescale.net:8181/16872
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Tested-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>