pwm: fsl-ftm: Convert to direct regmap API usage
The regmap core supports different endian modes for devices. This patch
convert to direct regmap API usage, preparing to support big endianness
for LS1 SoC.

Using the regmap framework it will be easy to support devices that only
differ in endianness with the same device driver.

Signed-off-by: Xiubo Li <Li.Xiubo@freescale.com>
Signed-off-by: Thierry Reding <thierry.reding@gmail.com>

---
This patch is pulled back from upstream:
commit 42fa98a9c3609c1aff466cb847e421c611cc9157

Change-Id: If7905b70eed8296dffa7ebde0dc03e951fcb7536
Reviewed-on: http://git.am.freescale.net:8181/19866
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>