pwm: fsl-ftm: Document 'big-endian' property
The same FTM PWM device can have a different endianness on different
SoCs. The device tree provides a property to describing this so that an
operating system device driver can handle all variants of the device.
Refer to the table below for the endianness of the FTM PWM block as
integrated into the existing SoCs:

	SoC     | FTM-PWM endianness
	--------+-------------------
	Vybrid  | LE
	LS1     | BE
	LS2     | LE

Signed-off-by: Xiubo Li <Li.Xiubo@freescale.com>
Signed-off-by: Thierry Reding <thierry.reding@gmail.com>

---
This patch is pulled back from upstream:
commit a535e2e0debc2255fcf60a11d73fbb0534454cc3

Change-Id: Icf9f1efe7a4fd121cb5568f8552c01043110b108
Reviewed-on: http://git.am.freescale.net:8181/19867
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>