dpa_offload: Avoid dereferencing an invalid pointer
In function "dpa_classif_mcast_create_group" from the dpa_classifier
driver there is a danger of dereferencing pgroup->entries pointer which
can be NULL in case of an error. To avoid this, in case of error this
pointer is checked before using it to release resources. Also, in
order to keep consistency of pgroup->entries in case of an allocation
error of pgroup->member_ids, the allocation of this structure and the
allocation of pgroup->member_ids have swicthed places.

Change-Id: I3d7ec1b01983a9c917f54d596e35afd8ab3539cc
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/18808
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>