dpa_offload: Remove dead code from "table_delete_entry_by_ref"
There is an unreachable code paragraph in function "table_delete_entry_by_ref".
The function doesn't work for indexed tables. In the first half, an error
is issued in this case. However, in the second half, another code paragraph
is testing again if the table type is "indexed". This branch will never
be reachable and was removed.

Change-Id: Ib200f49eafca8131d4c1696af2e61753ede41a3d
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/19135
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>