Documentation: fsl-quadspi: update the document
The patch updates the document by adding more information to describe the
DT proporties used by the Freescale Quadspi driver and the childs nodes.

For the child node for SPI NOR flash, we add the required property
("spi-max-frequency"), and refer to spi-nor-flash.txt for the optional
properties.

Signed-off-by: Huang Shijie <b32955@freescale.com>
The upstream status of this patch: http://patchwork.ozlabs.org/patch/343229/
Change-Id: Ib625035f0fd863840eca9dce66699836883bdfe3
Reviewed-on: http://git.am.freescale.net:8181/20128
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>
Tested-by: Zhengxiong Jin <Jason.Jin@freescale.com>