mtd: fsl-quadspi: use the information stored in spi-nor{}
We can get the read/write/erase opcode from the spi nor framework now.
What's more is that we can get the correct dummy cycles.

This patch uses the information stored in the spi_nor{} to remove the
hardcode in the fsl_qspi_init_lut().

Signed-off-by: Huang Shijie <b32955@freescale.com>
The upstream status of this patch can be found: https://patchwork.kernel.org/patch/4074971/
Change-Id: I32f982872df1729582f4122ac0dede934d749a04
Reviewed-on: http://git.am.freescale.net:8181/20129
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>
Tested-by: Zhengxiong Jin <Jason.Jin@freescale.com>