dpa_offload: Relocate dpa_classifier entry in index management list
For HASH tables sometimes when a user modifies the key of an entry
the entry needs to be moved to a different bucket. The index management
list is sorted so that indexes are always ascending. When tempering with
the position of an entry in the internal Cc nodes for a HASH table, the
position of the entry in the index management list must also be
updated.

Change-Id: If4b48282c407ed248dfd389cc0c8ae856fb7db75
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/23700
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-by: Matthew Weigel <Matthew.Weigel@freescale.com>