offline_port: convert dt into CPU endianness
The device tree values are stored in BE format by the dt compiler. In order to
support different architectures, not only BE, the values read by the offline
port driver should be converted into CPU endianness.

This patch convert the dt values into CPU endianness and fixes some sparse
errors related to endianness.

Signed-off-by: Rotariu Marian Cristian-B44860 <marian.rotariu@freescale.com>
Change-Id: Id7adce3b6f73619c58ac1cee9201bb63131f8d58
Reviewed-on: http://git.am.freescale.net:8181/21083
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/32496