dpa_offload: Fill in proper VSP parameters when policer is enabled
There are 2 different VSP profile parameters in the FMD API that need to
be filled in depending whether the policer is enabled or disabled. The
dpa_classifier was filling in the enqueue VSP profile all the time, but
this was used only when the next action was "enqueue". When the next
action is "policer" the policer VSP profile must be filled in.

Change-Id: I09e2a6fad7216e7ce7b3c1190a4441b7a9825705
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22064
Reviewed-by: Francois Massot-Pellet <Francois.Massot-Pellet@freescale.com>