dpa_offload: use SEC API for retrieving JR
SEC driver offers a function for geting the handle
to an available JR (caam_jr_alloc(...)
This is mainly needed for split-key calculation per
SA. There's no need to search for the JR in the
device tree. This patch removes the unneeded code
and replaces it with the proper SEC API call.

Change-Id: Ibc8b4b216d5699fe5ec7d3d49b8a3560ab79ede0
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/23860
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Reviewed-by: Matthew Weigel <Matthew.Weigel@freescale.com>