Fix cluster calculation issue for checking PCL10 state on e6500
Kernel could not get cluster information from device tree on PowerPC.
The previous way of cluster calculation could be wrong if all cores of
some clusters are disabled. For now PCL10 is only supported on e6500
cores. We hard coded the threads number of each cluster to ensure the
cluster calculation is right.

Signed-off-by: Jia Hongtao <hongtao.jia@freescale.com>
Change-Id: Ia329e48bac1a838828146df9b2caa7f527e329eb
Reviewed-on: http://git.am.freescale.net:8181/23177
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Yuantian Tang <yuantian.tang@freescale.com>
Reviewed-by: Zhuoyu Zhang <Zhuoyu.Zhang@freescale.com>
Reviewed-by: Yang Li <LeoLi@freescale.com>
Reviewed-by: Matthew Weigel <Matthew.Weigel@freescale.com>