This reverts commit 59d6bf6f6d2a2e99f2234ac572da5a5938fbd352
That commit tried to fix the section mismatch warning by moving the
clock driver struct to init section. This is definitely wrong because
the kernel would free the memories occupied by this struct after
boot while this driver is still registered in the driver core.
The kernel would panic when accessing this driver struct.

Signed-off-by: Tang Yuantian <Yuantian.Tang@freescale.com>
Change-Id: I3061188d09f504c9e1a0d8771a5336477404d016
Reviewed-on: http://git.am.freescale.net:8181/23314
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Matthew Weigel <Matthew.Weigel@freescale.com>