powerpc/fsl-soc: Disable CPC sole-data mode if KVM enabled
Set undocumented bit to disable CPC sole data mode, making it safe to
use the LRAT with KVM guests, which could otherwise hang the host
(under certain usage patterns) by accessing memory with the M bit
unset.  The guest still needs to set the M bit to operate properly, but
this removes a DoS vector.

Only do so if CONFIG_KVM is enabled as supposedly disabling this
affects performance, though I wasn't able to detect a difference
in the workloads I ran.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Change-Id: I5e69f03150f93556f3548ed4f8b3df65cfe6d802
Reviewed-on: http://git.am.freescale.net:8181/24843
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>
Reviewed-by: Matthew Weigel <Matthew.Weigel@freescale.com>