dpaa_eth: adjust advanced drivers to older kernel
Support for multiple drivers per kernel module comes in kernel 3.14
so we are forced to generate several modules for the advanced drivers.
This patch can be dropped when the code is rebased on a kernel newer
than 3.14. The problem is with the MODULE_DEVICE_TABLE macro that only
from 3.14 generates different named structured based on the name
parameter. The change that introduces that also needs a synchronized
change in scripts/mod/file2alias.c.

Signed-off-by: Madalin Bucur <madalin.bucur@freescale.com>