dpaa_eth: Add support for MACsec
This patch adds support for MACsec in DPAA.
The feature is implemented as a separate driver and hooks into the DPAA
Ethernet driver using the hooks exported by the latter.

The MACsec driver exposes a set of APIs allowing the user to configure
the hardware in MACsec mode and to change configuration at runtime (add/
remove secure channels, secure associations, change secure association
keys, etc.).

The following capabilities are supported by the current implementation:
	- MACsec enabled on multiple interfaces
	- one SecY per interface (MAC)
	- one TX Secure Channel per SecY
	- up to 16 RX Secure Channels per SecY
	- debugfs statistics (TX counters only) per CPU per interface

Signed-off-by: Camelia Groza <camelia.groza@freescale.com>
Signed-off-by: Andrei Vasiliu <andrei.vasiliu@freescale.com>
Signed-off-by: Iulia Manda <manda.iulia@freescale.com>
Signed-off-by: Cristian Sovaiala <cristian.sovaiala@freescale.com>
Change-Id: I8fe2883a716bf82dc14aae467f3e0eab140d236d
Reviewed-on: http://git.am.freescale.net:8181/33832
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Tested-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>