dpa_offload: Restore ONIC interface buffer pool
The ONIC interface only works when the offline port OH2 is using
buffer pool #17 which was prepared for it. This prevents the
unification of the DTS example files for the "ipsec_offload" and
"dpa_offload" demo applications, because the NF offloading demo
application requests the buffer pool #16 on the OH2 offline port
as NF IP Forwarding does not support VSPs.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I3e2996f6777e4b16aee1a3cf489f89b60a37f325
Reviewed-on: http://git.am.freescale.net:8181/36070
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Nicolae-Sebastian Grigore <sebastian.grigore@freescale.com>
Reviewed-by: Pinghua An <pinghua.an@freescale.com>
Reviewed-by: Honghua Yin <Hong-Hua.Yin@freescale.com>