powerpc/kvm: Disable in-kernel MPIC emulation for PREEMPT_RT_FULL
Current in-kernel MPIC emulation for PowerPC is unsuitable for RT. More
specifically, the MPIC lock becomes a sleeping mutex, causing concurrency issues
on a KVM guest enter. Converting the lock to a raw_spinlock is still not a
viable solution, since the current implementation  is prone to induce large
latencies with preemption and irqs disabled on the host kernel.

Patch aready applied in the upstream 3.18.13 RT patch, version 9 [1].

[1] http://www.spinics.net/lists/linux-rt-users/msg13340.html

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@freescale.com>
Change-Id: Ie942de00717527ea78c2471f6a5153489f432e89
Reviewed-on: http://git.am.freescale.net:8181/36530
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Scott Wood <scottwood@freescale.com>
Reviewed-by: Honghua Yin <Hong-Hua.Yin@freescale.com>